requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
PyMuPDF==1.22.5
python-docx==1.1.0
requests==2.31.0
scikit-learn==1.3.0
cloudinary==1.39.0
keybert==0.7.0
spacy==3.7.2
sentence-transformers==2.2.2
